Make component directory and add auto tag
