+++ src/components/MultiSelect.js
@@ -1,10 +1,10 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
-import scrollIntoViewIIHOC from './lib/scrollIntoViewIIHOC';
-import SingleSelectOption from './SingleSelectOption';
-const SingleSelectOptionHOC = scrollIntoViewIIHOC(SingleSelectOption);
+import scrollIntoViewIIHOC from '../lib/scrollIntoViewIIHOC';
+import MultiSelectOption from './MultiSelectOption';
+const MultiSelectOptionHOC = scrollIntoViewIIHOC(MultiSelectOption);
 
-export default class SingleSelect extends Component {
+export default class MultiSelect extends Component {
 
   render(){
     const {
@@ -14,22 +14,23 @@ export default class SingleSelect extends Component {
       disabled,
       isOptionsPanelOpen,
       isTouchDevice,
+      multiSelectSelectedIndexes,
+      multiSelectSelectedOptions,
       name,
-      nextPotentialSelectionIndex,
       options,
-      prefix,
-      singleSelectSelectedIndex,
-      singleSelectSelectedOption
+      nextPotentialSelectionIndex,
+      prefix
     } = this.props;
 
     return (
       <div
         className={`
           rrs__select-container
+          rrs__select-container--multiselect
           ${(disabled === true) ? 'rrs__select-container--disabled' : ''}
           ${(isTouchDevice === true) ? 'rrs__is-touch' : 'rrs__is-desktop'}
           ${(isOptionsPanelOpen === true) ? 'rrs__options-container--visible' : ''}
-          ${(altered) ? 'rrs__has-changed': ''}
+          ${altered ? 'rrs__has-changed': ''}
         `}
         role="listbox"
         tabIndex="0"
@@ -45,10 +46,14 @@ export default class SingleSelect extends Component {
         {!customLabelText &&
         <div className="rrs__label-container">
           <span className="rrs__label">
-            {prefix &&
-              <span>{prefix}</span>
-            }
-            {singleSelectSelectedOption.text}
+            <span className='rrs__multiselect__label'>
+              <span className='rrs__multiselect__label-text'>{`${prefix ? prefix + ' ' : ''}${multiSelectSelectedOptions.options[0].text}`}</span>
+              {multiSelectSelectedOptions.options.length > 1 &&
+              <span className='rrs__multiselect__label-badge'>
+                {`+ ${multiSelectSelectedOptions.options.length-1}`}
+              </span>
+              }
+            </span>
           </span>
           {caretIcon && caretIcon}
         </div>
@@ -60,14 +65,14 @@ export default class SingleSelect extends Component {
         >
           {options.length > 0 &&
             options.map((option, index) => (
-              <SingleSelectOptionHOC
+              <MultiSelectOptionHOC
                 scrollIntoViewScrollPaneRef={() => this.optionsContainer}
                 scrollIntoViewElementSelector={'rrs__option--next-selection'}
                 key={index}
                 index={index}
-                isTouchDevice={isTouchDevice}
                 option={option}
-                singleSelectSelectedIndex={singleSelectSelectedIndex}
+                isTouchDevice={isTouchDevice}
+                multiSelectSelectedIndexes={multiSelectSelectedIndexes}
                 nextPotentialSelectionIndex={nextPotentialSelectionIndex}
               />
             ))
@@ -75,7 +80,11 @@ export default class SingleSelect extends Component {
         </div>
 
         {name &&
-        <input type="hidden" name={name} value={singleSelectSelectedOption.value} />
+        <input
+          type="hidden"
+          name={name}
+          value={[multiSelectSelectedOptions.options.map(v => v.value)].join(',')}
+        />
         }
 
       </div>
@@ -83,7 +92,7 @@ export default class SingleSelect extends Component {
   }
 }
 
-SingleSelect.propTypes = {
+MultiSelect.propTypes = {
   altered: PropTypes.bool,
   caretIcon: PropTypes.oneOfType([
     PropTypes.string,
@@ -95,11 +104,22 @@ SingleSelect.propTypes = {
     PropTypes.element
   ]),
   disabled: PropTypes.bool,
-  singleSelectInitialIndex: PropTypes.number,
-  singleSelectSelectedIndex: PropTypes.number,
-  singleSelectSelectedOption: PropTypes.shape({
-    text: PropTypes.string,
-    value: PropTypes.string
+  multiSelectInitialSelectedIndexes: PropTypes.arrayOf(
+    PropTypes.number
+  ),
+  multiSelectSelectedIndexes: PropTypes.arrayOf(
+    PropTypes.number
+  ),
+  multiSelectSelectedOptions: PropTypes.shape({
+    altered: PropTypes.bool,
+    options: PropTypes.arrayOf(
+      PropTypes.shape({
+        name: PropTypes.string,
+        text: PropTypes.string,
+        value: PropTypes.string,
+        markup: PropTypes.object
+      })
+    )
   }),
   isTouchDevice: PropTypes.bool,
   isOptionsPanelOpen: PropTypes.bool,
@@ -112,6 +132,5 @@ SingleSelect.propTypes = {
       value: PropTypes.string.isRequired
     })
   ).isRequired,
-  prefix: PropTypes.string,
-  selectedValue: PropTypes.string
+  prefix: PropTypes.string
 };
